fix(ci): cargo-deny bitflags skip + 400-line guardrail compliance
Two CI gates the local pre-push checks missed (I ran 'deny check licenses'
only — CI runs the full 'check advisories bans licenses sources' — and
never ran scripts/check-file-size.sh):

Cargo Deny:
- portable-pty (#749) pins bitflags 1.x next to the workspace's 2.x; added
  the duplicate skip with the house-style reason

File Size Lint (11 files over 400 lines, all from this batch — three of
them were sitting at exactly 400 and one-line edits tipped them over):
- interactive.rs 494→384: types/traits/helpers split to
  interactive_types.rs (re-exported, call sites unchanged)
- interactive_tests.rs 573→294: resume/cancel/timeout/security/e2e tests
  split to interactive_session_tests.rs (mock infra shared via pub(super));
  the e2e insta snapshot moved with its module path
- diff_review.rs 548→275: draw layer split to diff_review_draw.rs, tests
  to diff_review_tests.rs
- interaction/mod.rs 467→386: draw_impl split to render.rs
- lifecycle.rs 404→269: MockTeardown/FakeClock/test accessors split to
  lifecycle_test_ports.rs (re-exported from lifecycle, imports unchanged)
- config/agents.rs 444→376: transport field tests moved to
  src/config/tests/agents_transport.rs
- prompt_input/tests/core.rs 402→354: #919 launch-option tests moved to
  tests/launch_options.rs
- keymap_tests.rs 403→400, codex.rs/opencode.rs/config/mod.rs 401→400:
  minimal line shaves (import-group merge / blank line / one redundant
  assert)

No behavior change: 5768 lib tests green before and after; fmt, the
clippy gate, full cargo-deny, and scripts/check-file-size.sh all pass
locally now.
