+++ src/config/tests.rs
@@ -7,6 +7,7 @@ use super::*;
 const MINIMAL_TOML: &str = "[project]\nrepo = \"owner/repo\"\n[sessions]\n[budget]\nper_session_usd = 5.0\ntotal_usd = 50.0\nalert_threshold_pct = 80\n[github]\n[notifications]\n";
 
 mod agents;
+mod agents_transport;
 mod behavior;
 mod configuration_doc;
 mod core;
